backend/requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pymongo==4.12.0
python-dotenv==1.1.0
requests==2.32.3
urllib3==2.4.0
Werkzeug==3.1.3
